requirements.txt
numpy<2.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyMuPDF
python-docx
pytesseract
Pillow
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1.tar.gz
nltk
Jinja2
WeasyPrint
reportlab
